refactor(prestige): extract shared flex primitives + minor cleanups
- New `Prestige/_shared.js`: palette, FALLBACK_AVATAR, PRESTIGE_CAP, and
  avatarBubble/levelBadge/bulletRow primitives. Lv50/Lv100/Lv100NoTrial CTAs
  and Status now consume the shared module instead of redefining each.
- pipeline.resolveUnconsumedPassedTrialName: parallelize the two
  independent listPassed + listConsumed reads.
- PrestigeService.prestige (awakening): unlock build-combo achievements
  in parallel instead of awaiting each one in sequence.
- Drop dead `consumedSet` (always empty) in PendingPrestigeBanner.
