+++ app/src/templates/application/Prestige/Lv50CTA.js
@@ -1,8 +1,10 @@
-const PRESTIGE_PURPLE = "#3B2A6B";
-const PRESTIGE_GOLD = "#F5C84B";
-const PRESTIGE_PURPLE_DEEP = "#2A1F4D";
-const PRESTIGE_LILAC = "#E5DFFF";
-const FALLBACK_AVATAR = "https://i.imgur.com/NMl4z2u.png";
+const {
+  PRESTIGE_PURPLE,
+  PRESTIGE_GOLD,
+  PRESTIGE_LILAC,
+  avatarBubble,
+  levelBadge,
+} = require("./_shared");
 
 function buildHeroBubble({ displayName, pictureUrl, level, liffUri }) {
   return {
@@ -52,26 +54,7 @@ function buildHeroBubble({ displayName, pictureUrl, level, liffUri }) {
           margin: "md",
           alignItems: "center",
           contents: [
-            {
-              type: "box",
-              layout: "vertical",
-              width: "36px",
-              height: "36px",
-              cornerRadius: "18px",
-              backgroundColor: PRESTIGE_PURPLE_DEEP,
-              borderColor: PRESTIGE_GOLD,
-              borderWidth: "1.5px",
-              flex: 0,
-              contents: [
-                {
-                  type: "image",
-                  url: pictureUrl || FALLBACK_AVATAR,
-                  size: "full",
-                  aspectMode: "cover",
-                  aspectRatio: "1:1",
-                },
-              ],
-            },
+            avatarBubble(pictureUrl),
             {
               type: "text",
               text: displayName,
@@ -81,29 +64,7 @@ function buildHeroBubble({ displayName, pictureUrl, level, liffUri }) {
               flex: 0,
               gravity: "center",
             },
-            {
-              type: "box",
-              layout: "vertical",
-              backgroundColor: PRESTIGE_PURPLE_DEEP,
-              cornerRadius: "10px",
-              paddingTop: "3px",
-              paddingBottom: "3px",
-              paddingStart: "10px",
-              paddingEnd: "10px",
-              borderColor: PRESTIGE_GOLD,
-              borderWidth: "1px",
-              flex: 0,
-              contents: [
-                {
-                  type: "text",
-                  text: `Lv.${level}`,
-                  color: PRESTIGE_GOLD,
-                  size: "xs",
-                  weight: "bold",
-                  align: "center",
-                },
-              ],
-            },
+            levelBadge(`Lv.${level}`),
           ],
         },
       ],
